loginView refactor
