+++ language_app/lib/components/inputs/auth_inputs/authInputStyle.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+TextStyle textStyle = TextStyle(fontFamily: 'Montserrat', fontSize: 20.0);
+EdgeInsets inputPadding = EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0);
