--- language_app/lib/views/default_view/defaultViewStyle.dart
@@ -1,5 +1,0 @@
-import 'package:flutter/material.dart';
-
-TextStyle textStyle = TextStyle(fontFamily: 'Montserrat', fontSize: 20.0);
-
-EdgeInsets inputPadding = EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0);
