--- language_app/lib/views/default_view/loginView.dart
@@ -1,1 +1,0 @@
-
